Integrate 555236a7b7e3c4d86afafc0f84d0c5a91384ccf3

+++ tensorboard/components/vz_example_viewer/vz-example-viewer.js
@@ -800,9 +800,9 @@ var vz_example_viewer;
             var jsonFeat;
             if (this.newFeatureType === INT_FEATURE_NAME) {
                 var valueList = [];
-                var ints = new FloatList();
+                var ints = new Int64List();
                 ints.setValueList(valueList);
-                feat.setInt64List(ints); // TODO(#2266): Preserves behavior.
+                feat.setInt64List(ints);
                 jsonFeat = { int64List: { value: valueList } };
             }
             else if (this.newFeatureType === FLOAT_FEATURE_NAME) {
